requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.6
djangorestframework==3.15.1
filelock==3.14.0
flatbuffers==24.3.25
fsspec==2024.5.0
gast==0.5.4
grpcio==1.64.0
grpcio-status==1.62.2
h5py==3.11.0
httplib2==0.22.0
idna==3.7
Jinja2==3.1.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
namex==0.0.8
networkx==3.3
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
requests==2.32.3
rich==13.7.1
rsa==4.9
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
termcolor==2.4.0
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
